requirements.txt
django==1.6.4
beautifulsoup4
pyyaml
pyatom
